pyproject.toml
[tool.poetry]
name = "essential-takehome"
version = "0.1.0"
description = ""
authors = ["Aaron Bannin <aaronbannin@Aarons-MacBook-Pro.local>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.9, !=3.9.7"
streamlit = "^1.28.2"
openai = "^1.3.5"
click = "^8.1.7"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"
jinja2 = "^3.1.2"
pandas = "^2.1.3"
langchain = "^0.0.341"
tabulate = "^0.9.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
